requirements.txt
undetected-chromedriver==3.5.5
beautifulsoup4==4.12.3
python-dotenv==1.0.1
requests==2.32.2
selenium==4.19.0
colorama==0.4.6
airtable-python-wrapper==0.15.3
ngrok==1.3.0